Fix Issue #17
Do a single GetProperty before to do the instrospection to avoid a
GetAllProperties to be performed first, this last do a UPnP Browse
action with an "*" filter which generates a warning with DLNA CTT
(TestCase 7.3.63.3 MM CDS Browse/Search Action).

This workaround is done for both MediaRenderer and MediaServer.

Fix issue https://github.com/01org/cloud-dleyna/issues/17